move improvise_rnn/utils.py -> utils.py for visibility
