+++ hw3/improvise_rnn/__init__.py
@@ -1,3 +1,2 @@
-from .utils import summary, cycle
 from .data_utils import load_music_utils
 from .qa import *
