+++ hw3/utils.py
